requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dronecan==1.0.26
empy==3.3.4
fastcrc==0.3.2
Flask==3.1.1
Flask-SocketIO==5.5.1
future==1.0.0
h11==0.16.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
MAVProxy==1.8.73
numpy
nunavut==1.9.0
opencv-python==4.12.0.88
pexpect==4.9.0
pillow==11.3.0
ptyprocess==0.7.0
pydsdl==1.22.2
pymavlink==2.4.48
pynmeagps==1.0.50
pyserial==3.5
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
setuptools==80.9.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
